fix(web): keep sign-in getServerSideProps off the client graph
Moving GSSP into SignInPage pulled multi-tenant-sso → shared/server →
dd-trace into the browser Turbopack graph. Host GSSP under auth/server
and re-export it from the pages shim only.
