+++ web/src/pages/auth/sign-in.tsx
@@ -1,8 +1,8 @@
 export {
   default,
-  getServerSideProps,
   SSOButtons,
   useHuggingFaceRedirect,
   FALLBACK_AUTH_PROVIDERS,
   type PageProps,
 } from "@/src/features/auth/SignInPage";
+export { getServerSideProps } from "@/src/features/auth/server/getSignInPageServerSideProps";
